+++ src/Element/TextareaElement.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * @link https://github.com/ixocreate
+ * @copyright IXOCREATE GmbH
+ * @license MIT License
+ */
+
+declare(strict_types=1);
+
+namespace Ixocreate\Schema\Element;
+
+use Ixocreate\Schema\Type\TypeInterface;
+
+final class TextareaElement extends AbstractSingleElement
+{
+    public function type(): string
+    {
+        return TypeInterface::TYPE_STRING;
+    }
+
+    public function inputType(): string
+    {
+        return 'textarea';
+    }
+
+    public static function serviceName(): string
+    {
+        return 'textarea';
+    }
+}