+++ src/SubSchemaReceiverInterface.php
@@ -9,6 +9,8 @@ declare(strict_types=1);
 
 namespace Ixocreate\Schema;
 
+use Ixocreate\Schema\Builder\BuilderInterface;
+
 interface SubSchemaReceiverInterface
 {
     public function receiveSchema(string $name, BuilderInterface $builder, array $options = []): SchemaInterface;