+++ src/Type/ActiveType.php
@@ -1,0 +1,45 @@
+<?php
+/**
+ * @link https://github.com/ixocreate
+ * @copyright IXOCREATE GmbH
+ * @license MIT License
+ */
+
+declare(strict_types=1);
+
+namespace Ixocreate\Schema\Type;
+
+use Doctrine\DBAL\Types\StringType;
+
+final class ActiveType extends AbstractType implements DatabaseTypeInterface
+{
+    /**
+     * @param $value
+     * @throws \Exception
+     * @return mixed
+     */
+    public function transform($value)
+    {
+        if (!\in_array($value, ['active', 'inactive'])) {
+            //TODO Exception
+            throw new \Exception("invalid type");
+        }
+
+        return $value;
+    }
+
+    public function convertToDatabaseValue()
+    {
+        return (string)$this;
+    }
+
+    public static function baseDatabaseType(): string
+    {
+        return StringType::class;
+    }
+
+    public static function serviceName(): string
+    {
+        return 'active';
+    }
+}