+++ src/Type/DatabaseTypeInterface.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * @link https://github.com/ixocreate
+ * @copyright IXOCREATE GmbH
+ * @license MIT License
+ */
+
+declare(strict_types=1);
+
+namespace Ixocreate\Schema\Type;
+
+interface DatabaseTypeInterface
+{
+    /**
+     * @return mixed
+     */
+    public function convertToDatabaseValue();
+
+    /**
+     * @return string
+     */
+    public static function baseDatabaseType(): string;
+}