+++ src/Type/Factory/SchemaTypeFactory.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * @link https://github.com/ixocreate
+ * @copyright IXOCREATE GmbH
+ * @license MIT License
+ */
+
+declare(strict_types=1);
+
+namespace Ixocreate\Schema\Type\Factory;
+
+use Ixocreate\Schema\Builder\Builder;
+use Ixocreate\Schema\Type\SchemaType;
+use Ixocreate\ServiceManager\FactoryInterface;
+use Ixocreate\ServiceManager\ServiceManagerInterface;
+
+final class SchemaTypeFactory implements FactoryInterface
+{
+    /**
+     * @param ServiceManagerInterface $container
+     * @param $requestedName
+     * @param array|null $options
+     * @return mixed
+     */
+    public function __invoke(ServiceManagerInterface $container, $requestedName, array $options = null)
+    {
+        return new SchemaType($container, $container->get(Builder::class));
+    }
+}