+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased](https://github.com/ixocreate/schema-package/compare/0.1.0...develop)
 
+### Changed
+- Merged with Type package, moved CMS and Media specific Elements to their respective packages
+
 ## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/schema-package/compare/master...0.1.0)
 
 ### Changed